+++ modules/queries/endpoll.js
@@ -20,13 +20,16 @@
  */
 
 exports.run = (Bastion, message, args) => {
-  message.delete().catch(e => {
-    Bastion.log.error(e.stack);
-  });
+  if (message.deletable) {
+    message.delete().catch(e => {
+      Bastion.log.error(e.stack);
+    });
+  }
 };
 
 exports.config = {
-  aliases: []
+  aliases: [],
+  enabled: true
 };
 
 exports.help = {
